+++ bloom_nofos/nofos/nofo.py
@@ -698,6 +698,7 @@ def find_broken_links(nofo):
             or tag.get("href", "").startswith("https://docs.google.com")
             or tag.get("href", "").startswith("#_heading")
             or tag.get("href", "").startswith("#_bookmark")
+            or tag.get("href", "") == "about:blank"
         )
 
     broken_links = []